Restructure the demo screen into per-category pages
The 2800-line _DemoScreenState held the state and UI of every demo
category in one class. DemoScreen is now a thin router that picks a
page per category (vision / speech-to-text / NLP / TTS / chat /
multimodal), each owning only its own state, inference glue and input
UI. Shared infrastructure moved into standalone classes composed by
the pages: DemoSession (run state, progress/error/result panels and
model downloads), CameraInput (device selection, preview, still
capture and realtime frame grabbing), WaveformController and the
transcript view.

Behavior is ported as-is apart from small cleanups: camera capture
failures surface through exceptions, the VLM page restores its sample
image when leaving the camera source, and the unused CameraFrame RGBA
buffer is gone.
